VA Assistaint/requirements.txt
Flask==2.3.3
twilio==9.0.4
python-dotenv==1.0.0
requests==2.31.0
pydub==0.25.1
huggingface-hub==0.20.0
flask-cors==4.0.0